requirements.txt
yt-dlp
ffmpeg-python
openai-whisper
transformers
torch